Problem: network parameters not reflected in the app hash (CRO-528)
Solution: grouped the current set of network parameters
into a enum; extended the app hash computation with it
+ modified the affected code
