+++ chain-core/src/lib.rs
@@ -21,25 +21,30 @@ pub mod tx;
 
 use blake2::Blake2s;
 use common::{hash256, MerkleTree, Timespec, H256};
+use init::params::NetworkParameters;
 use parity_scale_codec::{Decode, Encode};
 use state::RewardsPoolState;
 use tx::fee::Fee;
 
 /// computes the "global" application hash (used by Tendermint to check consistency + block replaying)
-/// currently: app_hash = blake2s(root of valid TX merkle tree || blake2s(scale bytes(rewards pool state)))
-/// TODO: it should include the fee policy / network parameters etc. once that becomes changeable
+/// currently: app_hash = blake2s(root of valid TX merkle tree
+/// || root of account/staked state trie || blake2s(scale bytes(rewards pool state)) || blake2s(scale bytes(network params)))
+/// TODO: cache (as many parts remain static)
 /// MUST/TODO: include node whitelists
 pub fn compute_app_hash(
     valid_tx_id_tree: &MerkleTree<H256>,
     account_state_root: &H256,
     reward_pool: &RewardsPoolState,
+    params: &NetworkParameters,
 ) -> H256 {
     let valid_tx_part = valid_tx_id_tree.root_hash();
     let rewards_pool_part = reward_pool.hash();
+    let network_params_part = params.hash();
     let mut bs = Vec::new();
     bs.extend(&valid_tx_part);
     bs.extend(&account_state_root[..]);
     bs.extend(&rewards_pool_part);
+    bs.extend(&network_params_part);
     hash256::<Blake2s>(&bs)
 }
